Do not apply timezone delta during feed read
The timezone delta right now is applied twice: during the read of the wall, and parsing of the wall. This, obviously causes post time to appear  shifted in future by exactly four hours. This patch tackles the issue by deciding to apply timezone delta only once: during parsing of the post, but not reading of the post.